I touched a lot of stuff trying to figure out how to get symbols working in MS equation fields.  I found a solution that works for Microsoft, but it breaks everything else.  Basically I changed direct.cfg so that it will emit something like {\f6\u-3910\'ab} when a rightarrow is needed.  This is great, but the unicode value is bogus and breaks all other rtf readers.
The right thing to do would be to emit {\f6\'ab} when not in a word equation field
and {\f6\u-3910\'ab} when in one.  I added such a routine in CmdMathChar() but did
not want to do all the spade work to move *all* the entries that use the symbol font
from direct.cfg.  

I think the best solution would be an even lower level routine putSymbolChar() that
would do the right thing.  This would still require moving the entries from the 
direct.cfg file.  putSymbolChar() would look like emit_math_char().



git-svn-id: https://svn.code.sf.net/p/latex2rtf/code/trunk@574 bedafb05-6a7b-466a-8c5e-1828fe372967